Use extern for SH functions
Add a dummy C implementation for those functions to satisfy the linker.
Note that the C implementation does not get included in the .bc file we pass to
SH. This is important for the nondet functions, otherwise SH will not replace
them with the right semantics.

Use the cc crate (together with the build script build.rs) to build and link the
C code.